% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanData.R
\name{CleanData}
\alias{CleanData}
\title{Title: create function cleandata}
\usage{
CleanData(data)
}
\arguments{
\item{data:}{is a data frame}
}
\value{
data
}
\description{
Title: create function cleandata
}
