% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation.R
\name{correlation}
\alias{correlation}
\title{Title correlation}
\usage{
correlation(data)
}
\arguments{
\item{data:data.frame}{}
}
\value{
p: scatterplot
       c: corrplot
}
\description{
Title correlation
}
