% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readdata.R
\name{readdata}
\alias{readdata}
\title{Title read data "african_crises(1).csv"}
\usage{
readdata(data)
}
\arguments{
\item{data:}{data.frame}
}
\value{
lire_data: data of african_crises(1).csv file
        s: summary
}
\description{
Title read data "african_crises(1).csv"
}
