% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegressionLineaire.R
\name{RegressionLineaire}
\alias{RegressionLineaire}
\title{Title linear_Model}
\usage{
RegressionLineaire(data)
}
\arguments{
\item{data1:}{is a data.frame}
}
\value{
list(scaterrplot,lm,summary,anova,stepAIC)
}
\description{
Title linear_Model
}
