% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadData.R
\name{ReadData}
\alias{ReadData}
\title{Title read data "african_crises.csv"}
\usage{
ReadData(data)
}
\arguments{
\item{data:}{is data.frame}
}
\value{
lire_data: data of african_crises.csv file
        s: summary
}
\description{
Title read data "african_crises.csv"
}
